+++ client/jetbrains/src/main/java/com/sourcegraph/config/ThemeUtil.java
@@ -68,7 +68,12 @@ public class ThemeUtil {
     @Nullable
     private static String getHexString(@Nullable Color color) {
         if (color != null) {
-            return "#" + Integer.toHexString(color.getRGB()).substring(2);
+            String colorString = Integer.toHexString(color.getRGB());
+            if (colorString.length() > 2) {
+                return "#" + colorString.substring(2);
+            } else {
+                return "#000000";
+            }
         } else {
             return null;
         }