+++ enterprise/cmd/worker/internal/permissions/bitbucket_projects.go
@@ -11,6 +11,7 @@ import (
 	"github.com/prometheus/client_golang/prometheus"
 
 	"github.com/sourcegraph/log"
+
 	"github.com/sourcegraph/sourcegraph/internal/errcode"
 
 	"github.com/sourcegraph/sourcegraph/cmd/frontend/globals"
@@ -81,6 +82,7 @@ type bitbucketProjectPermissionsHandler struct {
 
 // Handle implements the workerutil.Handler interface.
 func (h *bitbucketProjectPermissionsHandler) Handle(ctx context.Context, logger log.Logger, record workerutil.Record) (err error) {
+	logger = logger.Scoped("bitbucketProjectPermissionsHandler", "handles jobs to apply explicit permissions to all repositories of a Bitbucket Project")
 	defer func() {
 		if err != nil {
 			logger.Error("bitbucketProjectPermissionsHandler.Handle", log.Error(err))
@@ -92,7 +94,7 @@ func (h *bitbucketProjectPermissionsHandler) Handle(ctx context.Context, logger 
 	// get the external service
 	svc, err := h.db.ExternalServices().GetByID(ctx, job.ExternalServiceID)
 	if err != nil {
-		return errors.Wrapf(err, "failed to get external service %d", job.ExternalServiceID)
+		return errcode.MakeNonRetryable(errors.Wrapf(err, "failed to get external service %d", job.ExternalServiceID))
 	}
 
 	if svc.Kind != extsvc.KindBitbucketServer {
@@ -110,7 +112,7 @@ func (h *bitbucketProjectPermissionsHandler) Handle(ctx context.Context, logger 
 	}
 
 	if job.Unrestricted {
-		return h.setReposUnrestricted(ctx, repos, job.ProjectKey)
+		return h.setReposUnrestricted(ctx, logger, repos, job.ProjectKey)
 	}
 
 	repoIDs := make([]api.RepoID, len(repos))
@@ -118,7 +120,7 @@ func (h *bitbucketProjectPermissionsHandler) Handle(ctx context.Context, logger 
 		repoIDs[i] = api.RepoID(repo.ID)
 	}
 
-	err = h.setPermissionsForUsers(ctx, job.Permissions, repoIDs)
+	err = h.setPermissionsForUsers(ctx, logger, job.Permissions, repoIDs, job.ProjectKey)
 	if err != nil {
 		return errors.Wrapf(err, "failed to set permissions for Bitbucket Project %q", job.ProjectKey)
 	}
@@ -151,7 +153,7 @@ func (h *bitbucketProjectPermissionsHandler) getBitbucketClient(ctx context.Cont
 	return bitbucketserver.NewClient(svc.URN(), &c, cli)
 }
 
-func (h *bitbucketProjectPermissionsHandler) setReposUnrestricted(ctx context.Context, repos []*bitbucketserver.Repo, projectKey string) error {
+func (h *bitbucketProjectPermissionsHandler) setReposUnrestricted(ctx context.Context, logger log.Logger, repos []*bitbucketserver.Repo, projectKey string) error {
 	sort.Slice(repos, func(i, j int) bool {
 		return repos[i].ID < repos[j].ID
 	})
@@ -159,6 +161,8 @@ func (h *bitbucketProjectPermissionsHandler) setReposUnrestricted(ctx context.Co
 	for i, repo := range repos {
 		repoIDs[i] = int32(repo.ID)
 	}
+
+	logger.Info("Setting bitbucket repositories to unrestricted", log.String("project_key", projectKey))
 
 	err := h.db.Perms().SetRepoPermissionsUnrestricted(ctx, repoIDs, true)
 	if err != nil {
@@ -173,7 +177,7 @@ func (h *bitbucketProjectPermissionsHandler) setReposUnrestricted(ctx context.Co
 // Each repo is processed atomically. In case of error, the task fails but doesn't rollback the committed changes
 // done on previous repos. This is fine because when the task is retried, previous repos won't incur any
 // additional writes.
-func (h *bitbucketProjectPermissionsHandler) setPermissionsForUsers(ctx context.Context, perms []types.UserPermission, repoIDs []api.RepoID) error {
+func (h *bitbucketProjectPermissionsHandler) setPermissionsForUsers(ctx context.Context, logger log.Logger, perms []types.UserPermission, repoIDs []api.RepoID, projectKey string) error {
 	sort.Slice(perms, func(i, j int) bool {
 		return perms[i].BindID < perms[j].BindID
 	})
@@ -193,7 +197,7 @@ func (h *bitbucketProjectPermissionsHandler) setPermissionsForUsers(ctx context.
 	}
 
 	if len(mapping) == 0 {
-		return errors.Errorf("no users found for bind IDs: %v", bindIDs)
+		return errcode.MakeNonRetryable(errors.Errorf("no users found for bind IDs: %v", bindIDs))
 	}
 
 	userIDs := make(map[int32]struct{}, len(mapping))
@@ -209,6 +213,8 @@ func (h *bitbucketProjectPermissionsHandler) setPermissionsForUsers(ctx context.
 		}
 	}
 
+	logger.Info("Applying permissions to Bitbucket project repositories", log.String("project_key", projectKey))
+
 	// apply the permissions for each repo
 	for _, repoID := range repoIDs {
 		err = h.setRepoPermissions(ctx, repoID, perms, userIDs, pendingBindIDs)
@@ -223,7 +229,7 @@ func (h *bitbucketProjectPermissionsHandler) setPermissionsForUsers(ctx context.
 func (h *bitbucketProjectPermissionsHandler) setRepoPermissions(ctx context.Context, repoID api.RepoID, _ []types.UserPermission, userIDs map[int32]struct{}, pendingBindIDs []string) (err error) {
 	// Make sure the repo ID is valid.
 	if _, err := h.db.Repos().Get(ctx, repoID); err != nil {
-		return errors.Wrapf(err, "failed to query repo %d", repoID)
+		return errcode.MakeNonRetryable(errors.Wrapf(err, "failed to query repo %d", repoID))
 	}
 
 	p := authz.RepoPermissions{