+++ enterprise/cmd/worker/internal/permissions/bitbucket_projects_test.go
@@ -176,6 +176,7 @@ func TestSetPermissionsForUsers(t *testing.T) {
 	// set permissions for 3 users (2 existing, 1 pending) and 2 repos
 	err = h.setPermissionsForUsers(
 		ctx,
+		log.Scoped("test", "test"),
 		[]types.UserPermission{
 			{BindID: "pushpa@example.com", Permission: "read"},
 			{BindID: "igor@example.com", Permission: "read"},
@@ -185,6 +186,7 @@ func TestSetPermissionsForUsers(t *testing.T) {
 			1,
 			2,
 		},
+		"foo",
 	)
 	require.NoError(t, err)
 	check()
@@ -192,6 +194,7 @@ func TestSetPermissionsForUsers(t *testing.T) {
 	// run the same set of permissions again, shouldn't change anything
 	err = h.setPermissionsForUsers(
 		ctx,
+		log.Scoped("test", "test"),
 		[]types.UserPermission{
 			{BindID: "pushpa@example.com", Permission: "read"},
 			{BindID: "igor@example.com", Permission: "read"},
@@ -201,6 +204,7 @@ func TestSetPermissionsForUsers(t *testing.T) {
 			1,
 			2,
 		},
+		"foo",
 	)
 	require.NoError(t, err)
 	check()
@@ -208,6 +212,7 @@ func TestSetPermissionsForUsers(t *testing.T) {
 	// test with wrong bindids
 	err = h.setPermissionsForUsers(
 		ctx,
+		log.Scoped("test", "test"),
 		[]types.UserPermission{
 			{BindID: "pushpa", Permission: "read"},
 			{BindID: "igor", Permission: "read"},
@@ -217,6 +222,7 @@ func TestSetPermissionsForUsers(t *testing.T) {
 			1,
 			2,
 		},
+		"foo",
 	)
 	// should fail if the bind ids are wrong
 	require.Error(t, err)
@@ -228,6 +234,7 @@ func TestSetPermissionsForUsers(t *testing.T) {
 	// run the same set of permissions again
 	err = h.setPermissionsForUsers(
 		ctx,
+		log.Scoped("test", "test"),
 		[]types.UserPermission{
 			{BindID: "pushpa@example.com", Permission: "read"},
 			{BindID: "igor@example.com", Permission: "read"},
@@ -237,6 +244,7 @@ func TestSetPermissionsForUsers(t *testing.T) {
 			1,
 			2,
 		},
+		"foo",
 	)
 	require.NoError(t, err)
 	check()