+++ internal/database/feature_flags.go
@@ -508,18 +508,19 @@ func (f *featureFlagStore) GetUserFlags(ctx context.Context, userID int32) (map[
 	}
 
 	res := make(map[string]bool, len(flags))
+
 	for _, ff := range flags {
 		res[ff.Name] = ff.EvaluateForUser(userID)
+	}
 
-		// Org overrides are higher priority than default
-		for _, oo := range orgOverrides {
-			res[oo.FlagName] = oo.Value
-		}
+	// Org overrides are higher priority than default
+	for _, oo := range orgOverrides {
+		res[oo.FlagName] = oo.Value
+	}
 
-		// User overrides are higher priority than org overrides
-		for _, uo := range userOverrides {
-			res[uo.FlagName] = uo.Value
-		}
+	// User overrides are higher priority than org overrides
+	for _, uo := range userOverrides {
+		res[uo.FlagName] = uo.Value
 	}
 
 	return res, nil